Rename dump functions file
